Blog_Project/requirement.txt
fastapi
uvicorn
sqlalchemy
sqlalchemy[aysncio]
aiosqlite
python-dotenv
email-validator
psycopg2-binary